WebGoat Lesson 8 Solution
I use the WebGoat lesson structure in my Security+ courses.  WebGoat provides a good outline for basic security topics.  At every opportunity I dig deeper into each topic from a Security+ perspective, but WebGoat on its own does a great job of introducing the topics.  Lesson 8 has a challenge in it that I use to introduce the concept of vulnerability/security challenges to each of my Security+ students.  I place this discussion right after our discussion about PKI.  The first slide is used to Demo OpenSSL certificate creation and inspection.  The bottom of slide 2 has the challenge in it.  This challenge requires opening a Docker container, connecting and elevating your permissions and then decrypting a message.  It is a fun challenge.