OverTheWire Level 0-10 Solutions
https://overthewire.org/wargames/bandit/bandit0.html

These challenges start off pretty straight forward, but they force new students to use Linux.  In pentesting it is very important to know how to use Linux and since we live in a Windows world, this may be the first exposure a new student has to Linux.  For new pentest students, who are Windows users, I have them download and work through the Ubuntu App that they can get from the Microsoft store.  The Ubuntu App provides a Linux CLI with a good amount of capability.